+++ changelog.md
@@ -9,7 +9,10 @@
 * Out of season crops no longer break. They just don't grow
 * Florpium now generates much more but no longer drops a sh*t ton of nuggets with fortune
 * Players now have reduced block reach
-* Player's Ghost now also stores dropped XP
+* Player's Ghost 
+  * Are stronger but neutral
+  * Can now be right-clicked to get back items
+  * Now also stores dropped XP. Players now drop more experience on death
 * Creepers now drop 1-3 gunpowder instead of 0-2 and have a rare TNT drop
 * More advancements work
 * Chained copper armor can now be smelted/blasted